+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/array/ArBase.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.java.generics.resolver.support.array
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 15.12.2014
+ */
+interface ArBase<T, K, J> extends ArBaseLvl2<T[], List<T>> {
+
+}