+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/array/ArBaseLvl2.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.java.generics.resolver.support.array
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 15.12.2014
+ */
+interface ArBaseLvl2<T, K> {
+
+}