+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/array/ArRoot.groovy
@@ -1,0 +1,10 @@
+package ru.vyarus.java.generics.resolver.support.array
+
+import ru.vyarus.java.generics.resolver.support.Model
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 15.12.2014
+ */
+interface ArRoot extends ArBase<Model, ? extends Model, ? super Model> {
+}