+++ src/test/groovy/ru/vyarus/java/generics/resolver/support/wildcard/WCBaseLvl2.groovy
@@ -1,0 +1,9 @@
+package ru.vyarus.java.generics.resolver.support.wildcard
+
+/**
+ * @author Vyacheslav Rusakov 
+ * @since 15.12.2014
+ */
+interface WCBaseLvl2<T> {
+
+}