+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+* Add wildcards support
+* Improve complex generics resolution
+* Add access by generic name methods in GenericsContext
+* Add ability to disable cache using property and method to clear current cache
+
 ### 1.0.0 (2014-11-19)
 
 * Initial release